Unify CommMessages and ProtocolHandler to single helper object
