+++ casper/src/main/scala/coop/rchain/casper/util/comm/CasperPacketHandler.scala
@@ -20,7 +20,7 @@ import coop.rchain.comm.CommError.ErrorHandler
 import coop.rchain.comm.discovery.NodeDiscovery
 import coop.rchain.comm.protocol.routing.Packet
 import coop.rchain.comm.rp.Connect.{ConnectionsCell, RPConfAsk}
-import coop.rchain.comm.transport.CommMessages.packet
+import coop.rchain.comm.rp.ProtocolHelper.packet
 import coop.rchain.comm.transport.TransportLayer
 import coop.rchain.comm.{transport, PeerNode}
 import coop.rchain.metrics.Metrics
