--- comm/src/main/scala/coop/rchain/comm/ProtocolHelper.scala
@@ -1,29 +1,0 @@
-package coop.rchain.comm
-
-import com.google.protobuf.ByteString
-import com.google.protobuf.any.{Any => AnyProto}
-import coop.rchain.comm.protocol.routing._
-
-// TODO remove
-/**
-  * Utility functions for working with protocol buffers.
-  */
-object ProtocolHelper {
-
-  implicit def toProtocolBytes(x: String): ByteString =
-    com.google.protobuf.ByteString.copyFromUtf8(x)
-  implicit def toProtocolBytes(x: Array[Byte]): ByteString =
-    com.google.protobuf.ByteString.copyFrom(x)
-  implicit def toProtocolBytes(x: Seq[Byte]): ByteString =
-    com.google.protobuf.ByteString.copyFrom(x.toArray)
-
-  def toPeerNode(n: Node): PeerNode =
-    PeerNode(NodeIdentifier(n.id.toByteArray), Endpoint(n.host.toStringUtf8, n.tcpPort, n.udpPort))
-
-  def sender(proto: Protocol): Option[PeerNode] =
-    for {
-      h <- proto.header
-      s <- h.sender
-    } yield toPeerNode(s)
-
-}
