+++ comm/src/main/scala/coop/rchain/comm/transport/TcpTransportLayer.scala
@@ -4,7 +4,7 @@ import java.io.ByteArrayInputStream
 
 import coop.rchain.comm._, CommError._
 import coop.rchain.comm.protocol.routing._
-
+import coop.rchain.comm.rp.ProtocolHelper
 import cats._, cats.data._, cats.implicits._
 import coop.rchain.catscontrib._, Catscontrib._, ski._, TaskContrib._
 import coop.rchain.shared.{Cell, Log, LogSource}
